Allow per-request timeouts in NetworkClient

+++ clients/src/main/java/org/apache/kafka/clients/consumer/ConsumerConfig.java
@@ -299,9 +299,9 @@ public class ConsumerConfig extends AbstractConfig {
                                         Importance.LOW,
                                         CommonClientConfigs.RECONNECT_BACKOFF_MS_DOC)
                                 .define(RETRY_BACKOFF_MS_CONFIG,
-                                        Type.LONG,
-                                        100L,
-                                        atLeast(0L),
+                                        Type.INT,
+                                        100,
+                                        atLeast(0),
                                         Importance.LOW,
                                         CommonClientConfigs.RETRY_BACKOFF_MS_DOC)
                                 .define(AUTO_OFFSET_RESET_CONFIG,
